# Conflicts:
#	hyjf_am_server/hyjf_am_core/hyjf_am_config/src/main/java/com/hyjf/am/config/service/impl/BankConfigServiceImpl.java
#	hyjf_cs_server/hyjf_cs_trade/src/main/java/com/hyjf/cs/trade/config/WebMvcConfig.java
#	hyjf_platform_server/hyjf_admin/src/main/java/com/hyjf/admin/config/InterceptorConfig.java

+++ hyjf_am_server/hyjf_am_core/hyjf_am_market/src/main/java/com/hyjf/am/market/dao/model/auto/Ads.java
@@ -28,7 +28,7 @@ public class Ads implements Serializable {
 
     private String shareImage;
 
-    private Byte isIndex;
+    private Integer isIndex;
 
     private String startTime;
 
@@ -154,11 +154,11 @@ public class Ads implements Serializable {
         this.shareImage = shareImage == null ? null : shareImage.trim();
     }
 
-    public Byte getIsIndex() {
+    public Integer getIsIndex() {
         return isIndex;
     }
 
-    public void setIsIndex(Byte isIndex) {
+    public void setIsIndex(Integer isIndex) {
         this.isIndex = isIndex;
     }
